backend/requirements.txt
# ---- Core web ----
fastapi==0.110.0
uvicorn[standard]==0.29.0
dependency-injector==4.41.0

# ---- Upload / form ----
python-multipart==0.0.9

# ---- Azure services ----
azure-storage-blob==12.19.1
azure-storage-queue==12.9.0
azure-identity==1.16.0

# ---- Data & config ----
pydantic==2.6.3
pydantic[email]==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ---- Async / utils ----
httpx==0.27.0

# ---- Logging & retry ----
loguru==0.7.2
tenacity==8.2.3


sqlalchemy>=2.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# ---- Password hashing ----
bcrypt==3.2.2
passlib[bcrypt]==1.7.4

# ---- JWT ----
python-jose[cryptography]==3.4.0